sha2: refactor backends
